DMI_download.sh modified for new dir structure in DMI_icecharts

+++ validation/ice_charts/DMI_download.sh
@@ -3,36 +3,49 @@
 ME=`readlink -f $0`
 Vdir=`dirname $ME`
 
-cd /work/timill/DMI_tmp
-rm -r *
+tmpdir=/work/timill/DMI_tmp
+cd $tmpdir
 
-# download .imz files
-ftpdir=ftp://ftp.dmi.dk/icemar/wa/
-wget -r -nH -e robots=off --no-parent --cut-dirs=5 -R "index.html*" $ftpdir
+if [ 1 -eq 1 ]
+then
+   # download .imz files
+   rm -r *
+   ftpdir=ftp://ftp.dmi.dk/icemar/wa/
+   wget -r -nH -e robots=off --no-parent --cut-dirs=5 -R "index.html*" $ftpdir
+fi
 lst=*.imz
 
-mkdir -p tmp
-cd tmp
+Tmpdir=$tmpdir/tmp
+mkdir -p $Tmpdir
+cd $Tmpdir
 
 # unpack and cp to storage directory
 DMIdir=/work/shared/nersc/msc/DMI_icecharts/sigrid
 for g in ../*.imz
 do
+   pwd
    f=`basename $g`
    echo $f
    cyear=${f:0:4}
    unzip ../$f
 
+   sroot=${f%*.imz}
+
+   odir=$DMIdir/$cyear
+   mkdir -p $odir
+
+   Odir=$odir/$sroot
+   mkdir -p $Odir
+   echo mv * $Odir
+   mv * $Odir
+
    #########################################################################
-   # Launch script to extract MIZ
+   # Launch script to extract MIZ & calc MIZ width/area
    # - compares today's observation to relevant forecasts
    $Vdir/DMI_process.sh $cyear
-
-   # TODO plot on top of relevant FC?
-   # TODO calc MIZ area/width?
    #########################################################################
 
-   echo mv * $DMIdir/$cyear
-   mv * $DMIdir/$cyear
-done
+   #clean (only needed if testing)
+   rm -f $Tmpdir/*
 
+done
